requirements.txt
PyMySQL==1.0.2
pandas==1.5.3
requests==2.28.2
python-dotenv==0.21.1
cryptography==39.0.1